+++ lib/rspec/core.rb
@@ -105,6 +105,11 @@ module RSpec
   end
 
   module Core
+    # @private
+    # This avoids issues with reporting time caused by examples that
+    # change the value/meaning of Time.now without properly restoring
+    # it.
+    Time = ::Time.dup
   end
 
   def self.const_missing(name)