+++ lib/rspec/core/reporter.rb
@@ -38,7 +38,7 @@ module RSpec::Core
     end
 
     def start(expected_example_count)
-      @start = Time.now
+      @start = RSpec::Core::Time.now
       notify :start, expected_example_count
     end
 
@@ -89,7 +89,7 @@ module RSpec::Core
     alias_method :abort, :finish
 
     def stop
-      @duration = Time.now - @start if @start
+      @duration = RSpec::Core::Time.now - @start if @start
       notify :stop
     end