+++ Changelog.md
@@ -20,6 +20,8 @@ Enhancements
 * Add `format_docstrings { |str| }` config option. It can be used to
   apply formatting rules to example group and example docstrings.
   (Alex Tan)
+* Use dup of Time so reporting isn't clobbered by examples that modify Time
+  without properly restoring it.
 
 Bug fixes